+++ tests/puppeteer/campaigns/functional/BO/02_orders/01_orders/06_bulkUpdateOrdersStatus.js
@@ -13,7 +13,7 @@ const FOLoginPage = require('@pages/FO/login');
 const ProductPage = require('@pages/FO/product');
 const CartPage = require('@pages/FO/cart');
 const CheckoutPage = require('@pages/FO/checkout');
-const OrderConfirmationPage = require('@pages/FO/orderConfirmation');
+const OrderConfirmationPage = require('@pages/FO/checkout/orderConfirmation');
 // Importing data
 const {PaymentMethods} = require('@data/demo/paymentMethods');
 const {DefaultAccount} = require('@data/demo/customer');
