+++ tests/puppeteer/campaigns/sanity/06_checkoutFO/01_orderProduct.js
@@ -11,7 +11,7 @@ const HomePage = require('@pages/FO/home');
 const CartPage = require('@pages/FO/cart');
 const LoginPage = require('@pages/FO/login');
 const CheckoutPage = require('@pages/FO/checkout');
-const OrderConfirmationPage = require('@pages/FO/orderConfirmation');
+const OrderConfirmationPage = require('@pages/FO/checkout/orderConfirmation');
 const {DefaultAccount} = require('@data/demo/customer');
 const CartData = require('@data/FO/cart');
 const {PaymentMethods} = require('@data/demo/paymentMethods');
