+++ tests/puppeteer/pages/FO/category/index.js
