+++ tests/puppeteer/pages/FO/checkout/index.js
