+++ tests/puppeteer/pages/FO/cms/index.js
