+++ tests/puppeteer/pages/FO/product/index.js
