+++ tests/puppeteer/pages/install/index.js
